feat: turning vr-related components to typescript
