+++ package.json
@@ -59,6 +59,7 @@
 		"@egjs/common-demo": "github:naver/egjs#common-demo",
 		"@egjs/component": "^2.2.2",
 		"@egjs/visible": "^2.1.0",
+		"@types/webxr": "^0.1.0",
 		"babel-eslint": "^10.0.1",
 		"babel-loader": "^8.0.4",
 		"babel-plugin-add-module-exports": "^0.2.1",
